fix(d2-s4): files relocated to new folders
