+++ src/main/java/acme/entities/practicum/Session.java
@@ -1,4 +1,4 @@
-package acme.entities.course;
+package acme.entities.practicum;
 /*
  * Consumer.java
  *
