+++ src/main/java/acme/roles/Company.java
@@ -21,7 +21,7 @@ import javax.validation.constraints.NotNull;
 import org.hibernate.validator.constraints.Length;
 import org.hibernate.validator.constraints.URL;
 
-import acme.entities.course.Practicum;
+import acme.entities.practicum.Practicum;
 import acme.framework.data.AbstractRole;
 import lombok.Getter;
 import lombok.Setter;
